refactor: Modularize main.rs into separate command files
- Separate command implementations into src/commands/ module
- Create individual files for each command:
  - crawl.rs (296 lines): Article crawling functionality
  - index.rs (321 lines): OpenSearch indexing
  - ontology.rs (319 lines): Knowledge graph extraction
  - search.rs (83 lines): Vector search operations
  - serve.rs (515 lines): Server components (embedding, coordinator, distributed)
  - mod.rs (15 lines): Module exports

- Reduce main.rs from 1,937 to 443 lines (77% reduction)
- Keep main.rs focused on CLI definition and command routing
- Fix module imports to use baram:: instead of crate::

This improves:
- Code readability and maintainability
- Test isolation and coverage
- Module organization

Fixes #2
